Moved version5 changes to version4
