+++ src/reducers/migration.ts
@@ -4,7 +4,6 @@ import { migrateVersion0ToVersion1 } from './version1';
 import { migrateVersion1ToVersion2 } from './version2';
 import { migrateVersion2ToVersion3 } from './version3';
 import { migrateVersion3ToVersion4 } from './version4';
-import { migrateVersion4ToVersion5 } from './version5';
 import { currentAppStateVersion } from './AppState';
 
 export const migrateAppState: MigrationManifest = {
@@ -12,6 +11,5 @@ export const migrateAppState: MigrationManifest = {
     1: migrateVersion0ToVersion1,
     2: migrateVersion1ToVersion2,
     3: migrateVersion2ToVersion3,
-    4: migrateVersion3ToVersion4,
-    [currentAppStateVersion]: migrateVersion4ToVersion5,
+    [currentAppStateVersion]: migrateVersion3ToVersion4,
 };
